--- pymathics/graph/measures_and_metrics/__init__.py
@@ -1,6 +1,0 @@
-"""
-Graph Measures and Metrics
-
-Measures include basic measures, such as the number of vertices and edges, \
-connectivity, degree measures, centrality, and so on.
-"""
